Marksheet Systerm in java.
This is the group project that me and my friends had done in sem-1 project sumission. In this we have created a marksheet generation systerm with the basic knowledge of the java . In this code we have use concept like loop , switch case , if else , class and some basic concept of oops in java.